working on swing
