--- old-src/com/pmcguire/cmsc335/trafficsim/timer/TickListener.java
@@ -1,6 +1,0 @@
-package com.pmcguire.cmsc335.trafficsim.timer;
-
-public interface TickListener
-{
-    void handleTick(int ticks);
-}
