+++ src/main/java/com/andisolsoftware/trafficsim/light/State.java
@@ -1,0 +1,19 @@
+package com.andisolsoftware.trafficsim.light;
+
+/*
+Author:     Philip McGuire
+Date:       2021-10-01
+Class:      UMGC CMSC-335
+Project:    3
+ */
+
+/**
+ * State represents a stage in traffic lights
+ */
+public enum State
+{
+    GREEN,
+    YELLOW,
+    RED,
+    BLANK, // used for flashing
+}
