+++ src/main/java/com/andisolsoftware/trafficsim/timer/TickListener.java
@@ -1,0 +1,6 @@
+package com.andisolsoftware.trafficsim.timer;
+
+public interface TickListener
+{
+    void handleTick(int ticks);
+}
