+++ src/test/java/com/andisolsoftware/trafficsim/TrafficsimApplicationTests.java
@@ -1,0 +1,13 @@
+package com.andisolsoftware.trafficsim;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TrafficsimApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
